+++ atm-e2e-tests/test/pageobjects/topup/phone.page.js
@@ -1,7 +1,7 @@
 import BasePage, { tag } from '../base.page.js';
 
 class TopupPhonePage extends BasePage {
-    get screen()     { return tag('screen_topupPhoneInput') }
+    get screen()     { return tag('screen_getVerifyOtp') }
     get confirmBtn() { return tag('btn_ยืนยัน') }
     get cancelBtn()  { return tag('btn_ยกเลิก') }